Use explicit casting when needed
Turned on the `-Wconversion`, `-Werror`, `-Wfatal-errors` flags for GCC
builds. That should help avoid further issues with implicit type
conversions.

+++ src/Animation.h
@@ -92,7 +92,7 @@ public:
 	// which all belong to this animation.
 	// The render_offset is constant for all frames. The render_size is also
 	// the grid size.
-	void setupUncompressed(Point render_size, Point render_offset, int _position, int _frames, int _duration, unsigned short _maxkinds = 8);
+	void setupUncompressed(Point render_size, Point render_offset, unsigned short _position, unsigned short _frames, unsigned short _duration, unsigned short _maxkinds = 8);
 
 	void setup(unsigned short _frames, unsigned short _duration, unsigned short _maxkinds = 8);
