+++ src/MenuManager.cpp
@@ -1145,8 +1145,8 @@ void MenuManager::dragAndDropWithKeyboard() {
 		}
 		// clear power dragging if power slot was pressed twice
 		else if (slotClick == ACTIVATED) {
-			if (drag_power > 0 && pow->canUpgrade(pow->tablist.getCurrent())) {
-				pow->upgradePower(pow->tablist.getCurrent());
+			if (drag_power > 0 && pow->canUpgrade(static_cast<short>(pow->tablist.getCurrent()))) {
+				pow->upgradePower(static_cast<short>(pow->tablist.getCurrent()));
 			}
 			drag_src = 0;
 			drag_power = 0;
