+++ src/SaveLoad.cpp
@@ -262,8 +262,8 @@ void GameStatePlay::loadGame() {
 			else if (infile.key == "spawn") {
 				mapr->teleport_mapname = infile.nextValue();
 				if (mapr->teleport_mapname != "" && fileExists(mods->locate(mapr->teleport_mapname))) {
-					mapr->teleport_destination.x = toInt(infile.nextValue()) + 0.5f;
-					mapr->teleport_destination.y = toInt(infile.nextValue()) + 0.5f;
+					mapr->teleport_destination.x = static_cast<float>(toInt(infile.nextValue())) + 0.5f;
+					mapr->teleport_destination.y = static_cast<float>(toInt(infile.nextValue())) + 0.5f;
 					mapr->teleportation = true;
 					// prevent spawn.txt from putting us on the starting map
 					mapr->clearEvents();
